PID, bússola, testes do GPS

+++ pid.h
@@ -1,5 +1,5 @@
 #ifndef __PID_H__
 #define __PID_H__
-void pid_init(int new_p, int new_i, int new_d);
-int pid_update(int erro);
+void pid_init(double new_p, double new_i, double new_d);
+double pid_update(double erro);
 #endif
